replaced public tuples with named types

+++ Connectionist/Helper/SparseTrainingDataProvider.cs
@@ -10,12 +10,12 @@ namespace BrightWire.Connectionist.Helper
     internal class SparseTrainingDataProvider : ITrainingDataProvider
     {
         readonly ILinearAlgebraProvider _lap;
-        readonly IReadOnlyList<Tuple<Dictionary<int, float>, Dictionary<int, float>>> _data;
+        readonly IReadOnlyList<Tuple<Dictionary<uint, float>, Dictionary<uint, float>>> _data;
         readonly int _inputSize, _outputSize;
         IIndexableMatrix _inputCache = null, _outputCache = null;
         int _lastBatchSize = 0;
 
-        public SparseTrainingDataProvider(ILinearAlgebraProvider lap, IReadOnlyList<Tuple<Dictionary<int, float>, Dictionary<int, float>>> data, int inputSize, int outputSize)
+        public SparseTrainingDataProvider(ILinearAlgebraProvider lap, IReadOnlyList<Tuple<Dictionary<uint, float>, Dictionary<uint, float>>> data, int inputSize, int outputSize)
         {
             _lap = lap;
             _data = data;
@@ -27,7 +27,7 @@ namespace BrightWire.Connectionist.Helper
         public int InputSize { get { return _inputSize; } }
         public int OutputSize { get { return _outputSize; } }
 
-        public float Get(int row, int column)
+        public float Get(int row, uint column)
         {
             float ret;
             if (_data[row].Item1.TryGetValue(column, out ret))
@@ -35,7 +35,7 @@ namespace BrightWire.Connectionist.Helper
             return 0f;
         }
 
-        public float GetPrediction(int row, int column)
+        public float GetPrediction(int row, uint column)
         {
             float ret;
             if (_data[row].Item2.TryGetValue(column, out ret))
@@ -62,9 +62,9 @@ namespace BrightWire.Connectionist.Helper
             int index = 0;
             foreach (var row in rows) {
                 foreach (var item in _data[row].Item1)
-                    _inputCache[index, item.Key] = item.Value;
+                    _inputCache[index, (int)item.Key] = item.Value;
                 foreach (var item in _data[row].Item2)
-                    _outputCache[index, item.Key] = item.Value;
+                    _outputCache[index, (int)item.Key] = item.Value;
                 ++index;
             }
             return new MiniBatch(_inputCache, _outputCache);
